+++ CHANGELOG.md
@@ -4,6 +4,8 @@ Changelog
 **Unreleased**
 --------------
 
+- Fix fallback adapter support in IR code gen not recognizing Moshi parameters to primary constructors.
+
 0.27.0
 ------